Enhance onboarding modal and add theming
