+++ src/components/index.ts
@@ -1,0 +1,4 @@
+export { default as OnboardForm } from "./OnboardForm";
+export { default as Modal } from "./ui/modal";
+export * from "./ThemeProvider";
+export { default as Animated } from "./Animated";
